assay-python-sdk/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "assay"
version = "0.8.0"
description = "Python SDK for Assay - Formal quality testing for AI"
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
authors = [
  { name = "Assay", email = "hello@assay.dev" },
]
dependencies = [
  "PyYAML>=6.0",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "pytest-asyncio",
  "black",
  "isort",
  "mypy",
]
openai = ["openai>=1.0.0"]

[project.scripts]
assay = "assay.__main__:main"

[project.urls]
Homepage = "https://assay.dev"
Repository = "https://github.com/Rul1an/assay"
Issues = "https://github.com/Rul1an/assay/issues"

[tool.hatch.build.targets.wheel]
packages = ["assay"]

[tool.hatch.build.targets.sdist]
include = [
  "README.md",
  "assay/**",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
